Group_18/eventManagement/requirements.txt
aioredis==1.2.0
antiorm==1.2.1
asgiref==2.3.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
certifi==2019.3.9
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
daphne==2.2.5
db==0.1.1
db-sqlite3==0.0.1
Django==2.1.7
django-background-tasks==1.2.0
django-compat==1.0.15
django-crispy-forms==1.7.2
django-dbshell-plus==1.0.1
django-paypal==1.0.0
django-searchable-select==1.5.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.1
mysql-connector-python==8.0.15
mysqlclient==1.4.2.post1
Pillow==5.4.0
protobuf==3.7.1
PyHamcrest==1.9.0
pytz==2018.9
requests==2.21.0
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.1
zope.interface==4.6.0
